refactor: split PromptPreviewSection
